middleware/etcd: rename etcd to backend
This a smaller commit reduce the size of the PRs. It is not the
complete work.

Next steps:
- seprarate etcd and backend struct
- move etcd to it's own package
