+++ core/coredns.go
@@ -7,13 +7,13 @@ import (
 
 	// plug in the standard directives (sorted)
 	_ "github.com/coredns/coredns/middleware/auto"
+	_ "github.com/coredns/coredns/middleware/backend"
 	_ "github.com/coredns/coredns/middleware/bind"
 	_ "github.com/coredns/coredns/middleware/cache"
 	_ "github.com/coredns/coredns/middleware/chaos"
 	_ "github.com/coredns/coredns/middleware/dnssec"
 	_ "github.com/coredns/coredns/middleware/erratic"
 	_ "github.com/coredns/coredns/middleware/errors"
-	_ "github.com/coredns/coredns/middleware/etcd"
 	_ "github.com/coredns/coredns/middleware/file"
 	_ "github.com/coredns/coredns/middleware/health"
 	_ "github.com/coredns/coredns/middleware/kubernetes"
