+++ core/zmiddleware.go
@@ -5,6 +5,7 @@ package core
 import (
 	// Include all middleware.
 	_ "github.com/coredns/coredns/middleware/auto"
+	_ "github.com/coredns/coredns/middleware/backend"
 	_ "github.com/coredns/coredns/middleware/bind"
 	_ "github.com/coredns/coredns/middleware/cache"
 	_ "github.com/coredns/coredns/middleware/chaos"
@@ -12,7 +13,6 @@ import (
 	_ "github.com/coredns/coredns/middleware/dnssec"
 	_ "github.com/coredns/coredns/middleware/erratic"
 	_ "github.com/coredns/coredns/middleware/errors"
-	_ "github.com/coredns/coredns/middleware/etcd"
 	_ "github.com/coredns/coredns/middleware/file"
 	_ "github.com/coredns/coredns/middleware/health"
 	_ "github.com/coredns/coredns/middleware/hosts"
