+++ middleware.cfg
@@ -40,7 +40,7 @@
 180:file:file
 190:auto:auto
 200:secondary:secondary
-210:etcd:etcd
+210:etcd:backend
 220:proxy:proxy
 230:whoami:whoami
 240:erratic:erratic
