+++ middleware/backend/lookup_test.go
@@ -3,7 +3,7 @@
 package etcd
 
 import (
-	"github.com/coredns/coredns/middleware/etcd/msg"
+	"github.com/coredns/coredns/middleware/backend/msg"
 	"github.com/coredns/coredns/middleware/test"
 
 	"github.com/miekg/dns"
