+++ middleware/backend/stub.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/coredns/coredns/middleware/etcd/msg"
+	"github.com/coredns/coredns/middleware/backend/msg"
 	"github.com/coredns/coredns/middleware/proxy"
 
 	"github.com/miekg/dns"
