+++ middleware/kubernetes/kubernetes_test.go
@@ -10,7 +10,7 @@ import (
 	"k8s.io/client-go/1.5/pkg/api"
 
 	"github.com/coredns/coredns/middleware"
-	"github.com/coredns/coredns/middleware/etcd/msg"
+	"github.com/coredns/coredns/middleware/backend/msg"
 	"github.com/coredns/coredns/request"
 )
