+++ middleware/kubernetes/ns.go
@@ -4,7 +4,7 @@ import (
 	"net"
 	"strings"
 
-	"github.com/coredns/coredns/middleware/etcd/msg"
+	"github.com/coredns/coredns/middleware/backend/msg"
 
 	"github.com/miekg/dns"
 	"k8s.io/client-go/1.5/pkg/api"
