+++ test/etcd_cache_debug_test.go
@@ -7,7 +7,7 @@ import (
 	"log"
 	"testing"
 
-	"github.com/coredns/coredns/middleware/etcd/msg"
+	"github.com/coredns/coredns/middleware/backend/msg"
 	"github.com/coredns/coredns/middleware/proxy"
 	"github.com/coredns/coredns/middleware/test"
 	"github.com/coredns/coredns/request"
